+++ libc/src/time/nanosleep.h
@@ -18,5 +18,3 @@ int nanosleep(const struct timespec *req, struct timespec *rem);
 } // namespace __llvm_libc
 
 #endif // LLVM_LIBC_SRC_TIME_NANOSLEEP_H
-
-#include "include/time.h"